+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 5.2.0
 
 * Feat: Allow setting proguard via Options and/or external resources (#1728)
 * Feat: Add breadcrumbs for the Apollo integration (#1726)